day1_opencv/requirements.txt
opencv-python==4.9.0.80
gradio==4.44.0
numpy==1.26.4
matplotlib==3.8.4
pillow==10.3.0
